backend/requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
python-dotenv==0.15.0
elastic-apm==6.9.1
sqlalchemy==1.4.29
psycopg2==2.9.3

pydantic~=1.10.2
requests~=2.28.1
sqlmodel~=0.0.8
pip~=22.3.1
wheel~=0.37.1
Jinja2~=3.1.2
setuptools~=63.3.0
zipp~=3.11.0